Moved AdvancedHostResolver cache tests into separate test class. Skipping cache tests for NativeResolver and in travis-ci.
